oh man, wtf is wrong with go
so what happened was, the new function finishCall() calls consume(),
which can return an error. but that meant that call(), which calls
finishCall(), must also return an error. But now unary() calls call() so
now unary() has to return an error, and then the thing that calls
unary() has to return an error, and it just keeps going up and up the
stack. So now everywhere that an error could be returned that we want to
pass up, we have to do the if err != nil { return nil, err } thing over
and over.

there has to be a better way to do this kind of error-handling in go,
but this is my first actual project in go so i don't actually know how
to do the better way.