It's all working now
The bug was that when i did a for loop:

    for (var i = 0; i < 3; i = i + 1) {
      print i;
    }

This would get rewritten to:

    {
      var i = 0;
      while (i < 3) {
        {
            print i;
        }
        i = i + 1;
      }
    }

Now, the i in i < 3 is in a different block (hence, scope) than the i in
i = i + 1.  However, in the for-loop version, both instances of i appear
on the same line. Since our Token struct only stores the *line* on which
a variable appears, not the column, this meant that the two instances of
i both had literally the *same* Token struct.  This meant that when we
tried to do the resolver storing/lookup stuff, something goofy would
happen (I'm not exactly sure what, since i didn't dig in deeper once i
knew what was happening; presumably, it would overwrite the other
instance of i, or fail to store it seeing it was already stored, or
something of that sort).  Indeed, the following program would work, even
though the initial for-loop above didn't:

    for (var i = 0;
         i < 3;
         i = i + 1) {
      print i;
    }

This is because the two i's are now on different lines!

So, what's the solution? I decided to switch everything over to
pointers. This way, each Token struct instance would have its own
address, so our resolver wouldn't get confused about the multiple
appearances of i on the same line.