Retrying Rust
I realized that trying to make my own version of Vec is too hard given
that this is the first time ever I am even using Rust. So I decided to
just go with Rust's own Vec, and not worry about memory management for
now (and Rust can do whatever RAII crap it wants to do in the
background). I will still follow along with the book in C as well, to
get practice with manual memory management.